Extract geoserver functions into own package
